% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{parse.cli.args}
\alias{parse.cli.args}
\title{Parse specific command-line parameters.}
\usage{
parse.cli.args()
}
\value{
A vector containing:
\describe{
\item{output.type}{The type name of the output device, e.g., "png".}
\item{output.type.func}{The function matching the
name of the output device. The device forces the use of
inches for resolution units, with a DPI of 96.}
\item{output.device}{The output device name.}
}
}
\description{
\verb{--output-type=<value>} The value of this parameter should be an
output device.
Defaults to \code{png}
}
\details{
\verb{--output-fd=<value>} The integer value of the file descriptor that
the output device will write to.
If not specified, \code{parse.cli.args} tries to get
the FD from the \code{IMAGE_DATA_FD} envvar.
If the envvar is not specified, this defaults
to STDOUT.
}
