% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipc.R
\name{parse.value}
\alias{parse.value}
\title{Attempt to parse an R value from a Python value.
Warning: this method calls \code{eval}. Do no use with
untrusted data.}
\usage{
parse.value(x)
}
\arguments{
\item{x}{Any string matching R syntax representing a value.
Some examples are:
* "c('a','b')"
* "1"
* "list(1,2,3)"}
}
\value{
The parsed value as the proper R type.
For example, the R string \code{"c('a','b')"} becomes
the R vector \code{c('a','b')}
}
\description{
Attempt to parse an R value from a Python value.
Warning: this method calls \code{eval}. Do no use with
untrusted data.
}
