% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{partial}
\alias{partial}
\title{Partially apply parameter values to a method
and get a new method.}
\usage{
partial(func, ...)
}
\arguments{
\item{func}{The function to apply parameters to.}

\item{...}{Any named parameters and their values to apply to the function.}
}
\value{
A function whose parameters include the parameters from \code{func}
whose named parameters match those in \code{...}.
}
\description{
Partially apply parameter values to a method
and get a new method.
}
