% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipc.R
\name{read.csv.as.method.parameters}
\alias{read.csv.as.method.parameters}
\title{Parse a CSV containing a named list that can be applied
to a function whose parameters match the values in the named list.
Parameters are deserialized by calling \code{parse.value} on each value.}
\usage{
read.csv.as.method.parameters(parameter.csv)
}
\value{
A named list whose names are the first row of the CSV
and whose values are the second row of the CSV, matching
the same column as the names.
}
\description{
Parse a CSV containing a named list that can be applied
to a function whose parameters match the values in the named list.
Parameters are deserialized by calling \code{parse.value} on each value.
}
\examples{
csv <- "a,b,c\n1,2,3"
read.csv.as.method.parameters(csv)
# $a
# [1] 1
# $b
# [1] 2
# $c
# [1] 3
}
