% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipc.R
\name{read.dataframe}
\alias{read.dataframe}
\title{Parse a CSV or TSV into a dataframe.}
\usage{
read.dataframe(con = NULL, header = TRUE, na.strings = "NA", ...)
}
\value{
A dataframe.
}
\description{
Parse a CSV or TSV into a dataframe.
}
