% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipc.R
\name{read.method.parameters}
\alias{read.method.parameters}
\title{Parse a CSV containing a named list that can be applied
to a function whose parameters match the values in the named list.
This method executes and returns \code{read.csv.as.method.parameters}
after reading the CSV parameters from \code{con} or the device name
from the envvar \code{METHOD_ARG_READ_FD}.}
\usage{
read.method.parameters(con = NULL)
}
\arguments{
\item{con}{The connection to read CSV parameters from.}
}
\value{
The same as \code{read.csv.as.method.parameters}.
}
\description{
Parse a CSV containing a named list that can be applied
to a function whose parameters match the values in the named list.
This method executes and returns \code{read.csv.as.method.parameters}
after reading the CSV parameters from \code{con} or the device name
from the envvar \code{METHOD_ARG_READ_FD}.
}
